Moved hierarchies to meta/
