+++ include/boost/simd/arch/common/generic/function/autosplat.hpp
@@ -13,7 +13,7 @@
 
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/dispatch/hierarchy/functions.hpp>
-#include <boost/simd/sdk/hierarchy/simd.hpp>
+#include <boost/simd/meta/hierarchy/simd.hpp>
 #include <boost/simd/arch/common/tags.hpp>
 #include <boost/config.hpp>
