+++ include/boost/simd/arch/common/simd/function/extract.hpp
@@ -14,8 +14,8 @@
 #include <boost/simd/function/bitwise_cast.hpp>
 #include <boost/simd/detail/aliasing.hpp>
 #include <boost/simd/meta/as_arithmetic.hpp>
-#include <boost/simd/sdk/hierarchy/simd.hpp>
-#include <boost/simd/sdk/hierarchy/logical.hpp>
+#include <boost/simd/meta/hierarchy/simd.hpp>
+#include <boost/simd/meta/hierarchy/logical.hpp>
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/dispatch/adapted/std/integral_constant.hpp>
 #include <boost/config.hpp>
