+++ include/boost/simd/arch/common/simd/function/genmask.hpp
@@ -13,8 +13,8 @@
 #define BOOST_SIMD_ARCH_COMMON_SIMD_FUNCTION_GENMASK_HPP_INCLUDED
 
 #include <boost/simd/function/bitwise_cast.hpp>
-#include <boost/simd/sdk/hierarchy/logical.hpp>
-#include <boost/simd/sdk/hierarchy/simd.hpp>
+#include <boost/simd/meta/hierarchy/logical.hpp>
+#include <boost/simd/meta/hierarchy/simd.hpp>
 #include <boost/simd/meta/as_arithmetic.hpp>
 #include <boost/simd/constant/allbits.hpp>
 #include <boost/simd/constant/zero.hpp>
