+++ include/boost/simd/arch/common/simd/function/insert.hpp
@@ -13,8 +13,8 @@
 
 #include <boost/simd/function/genmask.hpp>
 #include <boost/simd/detail/aliasing.hpp>
-#include <boost/simd/sdk/hierarchy/simd.hpp>
-#include <boost/simd/sdk/hierarchy/logical.hpp>
+#include <boost/simd/meta/hierarchy/simd.hpp>
+#include <boost/simd/meta/hierarchy/logical.hpp>
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/config.hpp>
