+++ include/boost/simd/arch/x86/sse1/simd/function/make.hpp
@@ -11,7 +11,7 @@
 #ifndef BOOST_SIMD_ARCH_X86_SSE_SIMD_FUNCTION_MAKE_HPP_INCLUDED
 #define BOOST_SIMD_ARCH_X86_SSE_SIMD_FUNCTION_MAKE_HPP_INCLUDED
 
-#include <boost/simd/sdk/hierarchy/simd.hpp>
+#include <boost/simd/meta/hierarchy/simd.hpp>
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/config.hpp>
