+++ include/boost/simd/arch/x86/sse2/simd/function/aligned_store.hpp
@@ -14,7 +14,7 @@
 #include <boost/dispatch/adapted/common/pointer.hpp>
 #include <boost/align/is_aligned.hpp>
 #include <boost/simd/detail/aliasing.hpp>
-#include <boost/simd/sdk/hierarchy/simd.hpp>
+#include <boost/simd/meta/hierarchy/simd.hpp>
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/config.hpp>
