+++ include/boost/simd/arch/x86/sse2/simd/function/insert.hpp
@@ -12,7 +12,7 @@
 #define BOOST_SIMD_ARCH_X86_SSE2_SIMD_FUNCTION_INSERT_HPP_INCLUDED
 
 #include <boost/simd/detail/aliasing.hpp>
-#include <boost/simd/sdk/hierarchy/simd.hpp>
+#include <boost/simd/meta/hierarchy/simd.hpp>
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/config.hpp>
