+++ include/boost/simd/arch/x86/sse2/simd/function/load.hpp
@@ -12,7 +12,7 @@
 #define BOOST_SIMD_ARCH_X86_SSE2_SIMD_FUNCTION_LOAD_HPP_INCLUDED
 
 #include <boost/simd/function/make.hpp>
-#include <boost/simd/sdk/hierarchy/simd.hpp>
+#include <boost/simd/meta/hierarchy/simd.hpp>
 #include <boost/dispatch/function/overload.hpp>
 #include <boost/dispatch/adapted/common/pointer.hpp>
 #include <boost/config.hpp>
