+++ include/boost/simd/logical.hpp
@@ -108,7 +108,7 @@ namespace boost { namespace simd
   }
 } }
 
-#include <boost/simd/sdk/hierarchy/logical.hpp>
+#include <boost/simd/meta/hierarchy/logical.hpp>
 #include <boost/simd/detail/logical.hpp>
 
 #endif
